requirements.txt
beautifulsoup4~=4.10.0
requests~=2.28.2
pandas~=1.3.4
selenium~=4.0.0
pymongo~=3.12.1
py-object-pool~=1.1
pdfminer.six

DateTime~=4.4
urllib3~=1.26.7
polling2~=0.5.0
dateparser~=1.1.0
parse~=1.19.0
camelot-py~=0.9.0
PyPDF2~=2.12.1
opencv-python~=4.8.0.76
python3-ghostscript~=0.5.0
iteration-utilities~=0.12.1